Change 20080511-hqm-I by hqm@badtzmaru.home on 2008-05-11 08:39:39 EDT     in /Users/hqm/openlaszlo/trunk5/WEB-INF/lps/lfc     for http://svn.openlaszlo.org/openlaszlo/trunk/WEB-INF/lps/lfc
Summary: fix to make LzDatapointer.__LZgetNodes work in swf9

New Features:

Bugs Fixed:

Technical Reviewer: max
QA Reviewer: ptw
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

swf9 gets an error when the code in lzDatapointer tries to look for a property on 
a String object. The code needs to check whether the object it is dereferencing is
a hash table or not. 

This change verified that the object is not a String, assuming that if not, it must
be a hash table. Can POSNEXT ever be something besides a hash table or a String?


-      if ( posnext && posnext.pred == "range" ){
+      if ( posnext && !(posnext is String) && posnext['pred'] == "range" ){


    

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9068 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c