Change 20080517-lou-K by lou@loumac.local on 2008-05-17 13:20:14 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: clarify components->slider example

Bugs Fixed: LPP-4662

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Tests: As stated in a previous comment, I don't think this is a bug. Setting Value
higher that maxValue doesn't make sense, so you can't do it. There was an error in 
the original description of this bug:

"maxValue and Value change to 100"

Actually, Value is set to maxValue, and maxValue is unchanged. That sure seems right
to me.

I added this note to next to the Value box:

"(not greater than maxValue)"



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9216 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c