Change 20080613-lou-6 by lou@loumac.local on 2008-06-13 09:29:27 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: dguide. Chapter 18. Media Resources. fix spacing issues.

Technical Reviewer: (pending)
QA Reviewer: mwilde
Doc Reviewer: (pending)

Details: This problem has been found in a number of places now, and has to do with
a tag, like <sgmltag>, in the middle of a sentence with no space after it. Fixed for
this JIRA, and in several other places in this file.

Tests: visual verify



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9632 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c