Change 20080624-mwilde-6 by mwilde@mwilde-pc on 2008-06-24 13:12:22 PDT     in /home/mwilde/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fixed various reference guide pages

New Features:

Bugs Fixed:
LPP-6497
LPP-6481
LPP-6480
Partial fix for LPP-6478

Technical Reviewer: (pending)
QA Reviewer: liorio
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
Added the missing example to the LzURL page.

<rpc> page:
The reference to remotecall was a link in the old docs.

References to javarpc, soap, xmlrpc and remotecall were links in the old docs.

Links in the See Also section are fixed.

Shortdesc is no longer repeated after the See Also section.

<param> page:
The reference to remotecall was a link in the old docs.

<javarpc> page:
"javaRPC is a feature" should read "JavaRPC is a feature".

The See Also section is restored.

The shortdesc is no longer repeated at the end of the description.

<command> page:
The references to LzLazyReplicationManager were links in the old docs.

The keys attribute is listed as having a type of "[String]?". This should be "[String]".

The DisplayKeys attribute now has a type of "Dictionary".

The $lzc$set_key() function is now marked as private.

Tests: visual verify, test links



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9951 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c