Change 20080624-mwilde-1 by mwilde@mwilde-pc on 2008-06-24 16:54:14 PDT     in /home/mwilde/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fixed <baseform>, <basedatacombobox>, LzDataElement, LzDataText, and LzDataset reference guide pages.

New Features:

Bugs Fixed:
LPP-6466
LPP-6458
LPP-6457
LPP-6454
LPP-6449 (partial)

Technical Reviewer: (pending)
QA Reviewer: liorio
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
Fixed the following issues:

<baseform>:
3.4 has these tags highlighted as hyperlinks in the first paragraph and 4.x does not.

<submit> <checkbox> <combobox> <radiogroup> <form>

<basedatacombobox>:
Fixed the example.

LzDataText:
References to LzDataElement and LzDataNodeMixin are links in the old docs.

The description for the length attribute says that it is a read-only property, but it is in the read/write category.

Not a difference, but the reference to the W3C-specs should be converted to a link.

LzDataElement:
References to LzDataText and LzDataNodeMixin were links in the old docs.

DOCUMENT_NODE, ELEMENT_NODE, TEXT_NODE, and whitespaceChars were causing build warnings. I've added info to their descriptions to fill them out, and stop the warnings.

LzDataset:
The reference to <datapointer> was a link in the old docs.

"If a dataset is not explictly contained within in a ???" should read "If a dataset is not explictly contained within in a datasource" and datasource was a link in the old docs.

"the dataset's ??? event" should read "the dataset's ondata event", and ondata should link to the event.

querystring is documented below, so I removed it from the description.

All attributes are missing their tag types from the old documentation.

The acceptencodings, cacheable, getresponseheaders, nsprefix, and querystring attributes are marked as read-only in the old docs, but as read/write in the new ones. Is this correct?

cacheable attribute documents the defualt in the description, rather than in the default column as in the old docs.

clientcacheable says that, when true an argument will be added to prevent client caching. Is this true? It is opposite of the functionality of cacheable, so it's confusing.

While not a difference, it would be nice if the mulitrequest attribute listed its default in the table, instead of in the description.

nsprefix, proxied, and secureport are missing their defaults in the table from the old docs

getAllResponseHeaders() is missing its return documentation from the old docs.

getParams() describes its return object as an LzParams object, but it is actually of type LzParam.

Tests: visual verify, test links



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9959 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c