Change 20080625-hqm-J by hqm@badtzmaru.home on 2008-06-25 12:26:36 EDT     in /Users/hqm/openlaszlo/trunk4     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fix circular constraint example

New Features:

Bugs Fixed: LPP-6217

Technical Reviewer: lorio
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

+ We changed the way constraints work with states. When you apply a
state which applies a constraint on an attribute, it overwrites any
pre-existing constraint on that attribute. When the state is
un-applied, it removes the constraint, but does not attempt to restore
the original one.

I changed the example to use a slider to drag the squares around, thus
eliminating the use of the dragstate on the square, which would erase the
example constraints.

Tests:

example works now




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9982 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c