Change 20080625-mwilde-5 by mwilde@mwilde-pc on 2008-06-25 14:19:29 PDT     in /home/mwilde/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fixed various reference pages, and sorted protected members into public or private.

New Features:

Bugs Fixed:
LPP-6434 (partial)
LPP-6429 (partial)
LPP-6430
LPP-6425
LPP-6410
LPP-6528
LPP-6531
LPP-6530
LPP-6529
LPP-5958
LPP-5957
LPP-5951

Technical Reviewer: (pending)
QA Reviewer: liorio and amuntz
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
Fixed the following for LzDatapointer:
Reference to LzDataset is a link in the old docs.
Why is the attribute "xpath" listed as private (and thus is undocumented) if it's used in the example?
dupePointer(): "nullxpath and a falsererunxpath attribute" should be "null xpath and a false rerunxpath attribute"
xpathQuery(): reference to LzDataElement was a link in the old docs.

Fixed the following for LzDatapath:
LzDatapointer reference was a link in the old docs.
Since the Data Replication section became it's own chapter, the link to the pooling section under the pooling attribute points to the wrong place.
The link to the Data Replication section is now incorrect because it became its own chapter in the replication attribute.
updateData(): reference to LzInputText is a link in the description in the old docs.

Restored links to other components on the <frame> page.

Restored the missing example on the LzAudio page.

Restored the links to other components on the <u> page.

Fixed compiler warnings in LzDataNode and LzDataRequest for attributes without categories.

Sorted protected attributes and methods into public and private.
    
Tests: visual verify, test links



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@9999 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c