Change 20080625-mwilde-0 by mwilde@mwilde-pc on 2008-06-25 17:25:55 PDT     in /home/mwilde/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fixed various reference guide pages.

New Features:

Bugs Fixed:
LPP-6356 (partial)
LPP-6369
LPP-6373
LPP-6377
LPP-6388 (partial)
LPP-6348
LPP-6358
LPP-6361
LPP-6363
LPP-6368
LPP-6390


Technical Reviewer: (pending)
QA Reviewer: llundrigan
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

<view>:
LzNode is no longer a link.
clickregion attribute description: "befure" should be "before"
font attribute description: <text> and <inputtext> tag references are improperly formatted and have "@c{" before and "}" after them.
frame attribute description: version 3.4 had a much more detailed description. Should this be ported over, or has the functionality of this attribute changed?
opacity attribute: 1.0 and 0.0 have the same @c{ formatting mistake.
proxyURL attribute: "If the value is a string, treat it a URL to the proxy server." should read "If the value is a string, treat it as URL to the proxy server. "
subviews attribute: Missing type. 3.4 lists it as "Array of LzView"
getContextMenu() : While not a difference, it would look better if there was not a repeat of the function name before its description.
getMouse(): While not a difference, it'd be nice to put "string" in the xory argument type, instead of saying it in the description.
init(): "See LzNode.init() for more" has been changed to "See init" for more. The old wording was clearer.
setAADescription(): While not a difference, the description of its argument is the same as the description of the argument of setAAName(). Should it read "Sets the accessibility description for this view" instead?
setContextMenu(): While not a difference, it would look better if there was not a repeat of the function name before its description.
The <view> link under the Structure category used to go to a page named <view> that is now called <splashview>. It now goes to the <view> (LzView) page. This is significant, because a view in the splash context has subset of the features in <view>.

<import>:
<library> tag reference is a link in the 3.4 docs.
The onerror, onload, and ontimeout attributes are marked as type LzScript in the new docs. They're of type LzDelegate in the old docs. Also, should these be documented as events?

LzContextMenu:
Not a difference, but all of the functions repeat the name of the function in the description. It would look better if they didn't.
hideBuiltInItems() is missing its description from the old docs.
Not a difference, but should onMenuOpen be documented as an event?

<drawview>:
In "drawview extends LzView", "DrawView" should be capitialized.
globalAlpha attribute: In the description, "If the attribute is set to values outside this range, they are ignored." should read "If the attribute is set to values outside this range, it is ignored. " 

General:
Clicking on the link for the 4.x Latest trunk branch RingDing release creates a new tab with the heading "OpenLaszlo 4.2.x Reference", The title at the top of the page also shows this.
I believe it is supposed to say "OpenLaszlo 4.1.x Reference"
When ver. 3.4 of the refGuide opens it shows LZX Reference as the heading and browser title.

Added docs for <resizelayout>, <reverselayout>, <simpleboundslayout>, and <stableborderlayout>. 

<inputtext>:
Page is listed only as <inputtext> without (LzInputText).
<text> tag references are links in the old docs.

Tests: visual verify, test links



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@10006 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c