Change 20080820-hqm-k by hqm@badtzmaru.home on 2008-08-20 15:21:24 EDT     in /Users/hqm/openlaszlo/trunk/WEB-INF/lps/lfc     for http://svn.openlaszlo.org/openlaszlo/trunk/WEB-INF/lps/lfc
Summary: make swf9 debugger only print to  console  if $debug is set

New Features:

Bugs Fixed:

Technical Reviewer: pbr
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

Currently, in swf9, and call to Debug.write causes the debug console
field to become visible in the app, regardless of the setting of the debug flag.

It creates a problem in testing if this happens to obscure the application. 

This change conditionalizes the appearance of the console window to match the
debug flag.
    

Tests:

this should not show anything 
<canvas oninit="Debug.write('foo')"/>

this should show 'foo'
<canvas debug="true" oninit="Debug.write('foo')"/>




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@10726 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c