Change 20080823-hqm-9 by hqm@badtzmaru.home on 2008-08-23 00:36:37 EDT    in /Users/hqm/openlaszlo/trunk/WEB-INF/lps/lfc    for http://svn.openlaszlo.org/openlaszlo/trunk/WEB-INF/lps/lfc
Summary: implement cursor services for swf9

New Features:

Bugs Fixed: LPP-6853

Technical Reviewer: max
QA Reviewer: pbr
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

+ Implemented global "use hand cursor" behavior: added flag "LzMouseKernel.showhandcursor", and
added check for it in the LzSprite mouse event handler __mouseEvent().

+ Implemented global cursor in LzMouseKernel. This instantiates an LzSprite
with the specified resource. We might want to cache that sprite instead of instatiating
it each time.

+ Implemented LzSprite.setCursor

+ properly restore the user-defined global cursor set by the last call to LzMouseKernel.setCursorGlobal, if any,  when you move out of a view that has a custom cursor

NOTE: there is an issue with cursors which are PNG files and their
"hot spot". If the images are stored as PNG files, how do we specify
the the location of the "hot spot"??


NOTE: there is something not quite right about the double arrow cursor which is displayed
in the grid example "stretcher" controls; it appears to have it's left side clipped off.  This
seems to be a bug in the LzSprite bitmap conversion we are doing, to get around the
spurious mouseout events that happen with addChild of a swf resource





Tests:

test/cursors.lzx

examples/components/grid_example.lzx  when moving mouse over the column stretching
controls, the column stretcher cursor is active (double arrow)



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@10747 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c