Change 20080824-hqm-M by hqm@badtzmaru.home on 2008-08-24 10:26:21 EDT     in /Users/hqm/openlaszlo/trunk2/WEB-INF/lps/lfc     for http://svn.openlaszlo.org/openlaszlo/trunk/WEB-INF/lps/lfc
Summary: improvement for swf9 cursor

New Features:

Bugs Fixed:

Technical Reviewer: max
QA Reviewer: pbr
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
    
I wanted to use the actual resource for the cursor, instead of the bitmap that we build in LzSprite,
but if it's a swf, we get bogus mouseout events when the swf resource is displayed.

I found that if I make a container and put the resource in it, offset by one pixel vertically,
then the bogus mouse events don't get generated. 

So this ends up with the cursor off by one pixel vertically, but it doesn't generate spurious mouseout
and mouseover events.

I'm not sure if there's any way to compensate for this pixel offset at another layer of containment
or something. There really seemed to be a flash 9 bug with putting a swf resource at the same
location as the mouse pointer.

I wonder also if this approach could be made to work to obviate the need for the copyBitmap we're doing in
LzSprite resources ??


Tests:

test/cursors.lzx 
examples/components/grid_example.lzx swf9



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@10758 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c