Change 20080926-hqm-v by hqm@badtzmaru.home on 2008-09-26 13:05:30 EDT     in /Users/hqm/openlaszlo/trunk4     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fix for <import> loading timeout behavior

New Features:

Bugs Fixed: LPP-7062

Technical Reviewer: pbromanik 
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
    
+ Make LzLibraryLoader override LzLoader.returnData. 

When loading a library, the load operation doesn't implement the whole
data loader protocol; there is no data object returned in the case of
an error. There is only a successful load and a timeout condition.

In the case of a successful load, the library calls snippetLoaded as it's last operation.

In the case of a timeout, the LzLoadQueue calls returnData, so this patch 
overrides that to just send the ontimeout event to the library. 

There is still no way for an explicit onerror to happen now, because
the LPS server does not generate any error swf in the event that it
cannot find the library.


Tests:

+ A timeout could be acheived as reported in the bug report by loading the test case,
disabling the network, and then trying to load the library.


(For testing, I actually modified servlets/responders/ResponderLIB.java
temporarily to call Thread.sleep for 60 seconds, in order to create
timeout condition, to verify that ontimeout event is sent to library
object in example, so I didn't have to disable the network)




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@11250 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c