Change 20081001-hqm-S by hqm@badtzmaru.home on 2008-10-01 22:28:30 EDT     in /Users/hqm/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: combobox gets error in swf9

New Features:

Bugs Fixed: LPP-7057

Technical Reviewer: promanik
QA Reviewer: mamye
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
 
+ ensure that selection returned by selectionmanager points to a view

The baselist class tries to hilite the current selection, and if the
replication is 'lazy', the selection manager will return a
LzDatapointer instead of a view. This change just checks that the
thing being hilited is a view. There might be some better way to
ensure this, but it only seems to be an issue when the list is first
displayed.

Tests:

test case from bug report



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@11309 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c