Change 20081118-hqm-p by hqm@badtzmaru.home on 2008-11-18 21:42:12 EST     in /Users/hqm/openlaszlo/trunk4     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fixes for swf9 import libraries

New Features:

Bugs Fixed: LPP-7370, LPP-7639

Technical Reviewer: ptw
QA Reviewer: dda
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

+ When nodes with id's are declared in a <import> library, we need to
make a reference to the var in the code so that it gets compiled in
loaded by the flash loader.

We need to declare the global both in the main app and in the library,
because these are compiled as separate apps with separate intermediate
.as files in difference source code directories.

+ resources can now be embedded and accessed in the library .swf, and
should be compiled into the library, we don't need to do the swf7/8 hack
of compiling them into the main app.
    

Tests:
test/snippets/import-globals.lzx
test/snippets/import-resource.lzx




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@11827 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c