Change 20081209-lou-K by lou@loumac.local on 2008-12-09 08:53:28 AST     in /Users/lou/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: dguide: emphasize Debug.debug over Debug.write in examples

Bugs Fixed: LPP-7437 (partial)

Technical Reviewer: tucker
QA Reviewer: (pending)
Doc Reviewer: (pending)

Details: these are minor changes to the checkin described in the previous comment.

The change to the views chapter eliminates a reference to the debugger, which
in no longer used in the example.

Please review this and the changes in the previous comment.
    
Tests: visual verify



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@12022 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c