Change 20090121-bargull-Z64 by bargull@dell--p4--2-53 on 2009-01-21 23:42:19     in /home/Admin/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: fix tab focus bug in IE

New Features:

Bugs Fixed: LPP-7610

Technical Reviewer: max
QA Reviewer: promanik
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
When "tabEnabled" is set to false on TextFields, a tab-press focuses the next focusable object _outside_ of the flash-app. This is an IE-only issue and has been reported to Adobe. 
To workaround that issue, "tabEnabled" is set to true for IE-FlashPlayer environments and an event-listener for "FocusEvent.KEY_FOCUS_CHANGE" is installed to prevent the default behaviour and give lz.Focus the chance to control the focus handling.
(I also needed to change "LzInputTextSprite#gotBlur", it sets now the focus to the stage instead of null when removes the focus from a TextField.)

    

Tests:
testcase from bugreport



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@12596 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c