Change 20090209-hqm-Y by hqm@badtzmaru.home on 2009-02-09 20:54:26 EST     in /Users/hqm/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: add support for user comments in docs

New Features:

Bugs Fixed:

Technical Reviewer: liorio
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

+ Added new flag in docs/src/xsl/parameters.xsl, show.usercomments

 <!-- Add link for display/submission of user community comments -->
 <xsl:param name="show.usercomments">1</xsl:param>

When enabled, this adds in a user comments link at the bottom of each
doc page, via the common-html.xsl user.footer.content template.

The page docs/includes/commentwidget.js can be configured to point to
the server which supplies the comments. If the server is unavailable,
the comments will silently be omitted from the page.

Tests:

This is will be testable in the docs when the next nightly build of
the docs goes up on labs at
http://labs.openlaszlo.org/trunk-nightly/docs/reference/



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@12805 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c