Change 20090220-hqm-R by hqm@badtzmaru.home on 2009-02-20 01:18:46 EST     in /Users/hqm/openlaszlo/trunk3     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: support for swf9 incremental compilation

New Features:

Bugs Fixed:

Technical Reviewer: dda
QA Reviewer: ptw
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

This change allows the use of the swf9 flex compiler's 'incremental' compilation option.

The flex incremental compiler uses timestamps on the .as files, and a
cache db file of these timestamps, to determine which files to
recompile.

So this change makes the script compiler write the intermediate .as
files to a deterministically named directory
(tmp/lzswf9/build/path-to-app/) and tries to ensure that only files
whose content is actually changed are modified on disk.

For speed, this change adds an in-memory cache to the swf9 script
compiler of every intermediate .as file (which could get large I
imagine) to decide whether the content has changed on disk.


+ adds 'compiler.swf9.incremental'  flag to lps.properties

- This change doesn't add any way to pass the 'incremental' flag in
via query arg or command line args to the tag compiler. The only way
to set it right now is via the lps.properties file.

Tests:

compile a large test app with 300 view-classes (views with
constraints, that produce anonymous classes). The first compile takes
20 seconds, the second incremental compile takes 16 seconds.

About half the time is spent in the tag compiler and script compiler,
and the other half the time spent in the flex compiler (for the first
compile).

More specifically when I call just the flex compiler by hand, the
first compile takes about 9 seconds, and the second (incremental)
compile takes about 2 seconds.

So something is eating another three seconds, I think it is my code
which string-compares each cached javascript class file to determine
whether to rewrite the file on disk.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@12963 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c