Change 20090223-hqm-X by hqm@badtzmaru.home on 2009-02-23 01:00:02 EST     in /Users/hqm/openlaszlo/trunk3     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary:  fix swf9 compilation for windows

New Features:

Bugs Fixed: LPP-7790

Technical Reviewer: ptw
QA Reviewer: promanik
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:

The patch I made in r12963 to support swf9 incremental compilation needed to put all the
intermediate .as files for an app in a deterministically named directory. I was computing this
directory name by appending the Java working dir and the path to the application source file.

Unfortunately, on Windows, the path to the app source file can start with a disk drive specifier, like "C:"
which will make an illegal filename if it occurs in the middle of a pathname.


+ strip out windows disk drive prefixes, like "C:", from pathnames when building
the path to the temp working directory for compiling swf9 apps


Tests:

+ compile an app in swf9 on windows, without error
+ compile an app in swf9 on mac, without error




git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@13011 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c