Change 20090222-bargull-Tq7 by bargull@dell--p4--2-53 on 2009-02-22 23:04:30     in /home/Admin/src/svn/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: focus and select issue for lz.list

New Features:

Bugs Fixed: LPP-5159 (bad interaction of "baselistitem#_doMousedown" and "list#_doFocus"), LPP-1322 (List fails to properly select textlistitem after initial load)

Technical Reviewer: max
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:

Release Notes:

Details:
Whenever lz.list is focused, it'll ensure the current selection is visible by calling "list#ensureItemInView(..)". This may cause a bug, if
the list hasn't got focus and a user scrolls down the list (using the scrollbar does not give focus to the list!), because if a user now wan
ts to select a list-entry by clicking on it, the list gets focus and "ensureItemInView" will be called => list jumps back to the current se
lection (this is not the item the user clicked on, selection will actually happen at a later time).
To stop this behaviour, I've introduced a new flag on baselist ("__focusfromchild"), so it's possible to detect when a list-item set focus, in which case "list#ensureItemInView(..)" must not be called.
    

Tests:
testcases at LPP-5159, LPP-1322



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@13016 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c