Change 20090226-sebastianwagner-N by sebastianwagner@sebastian-wagners-macbook-pro.local on 2009-02-26 00:01:55 CET     in /Users/sebastianwagner/Documents/work/openlaszlo/branches/openlaszlo/trunk     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: javaRPC Custom Mappings 

New Features: apply custom mappings for javaRPC

Bugs Fixed: LPP-7788, LPP-7795

Technical Reviewer: (pending)
QA Reviewer: (pending)
Doc Reviewer: (pending)

Documentation:
The changes in the JavaDataSource are made in a way that it should not break the current system. The new type-mapping is only used if the RPC Class does extend the LaszloRPCAdapter.

The Custom Type mappings only does lookup if the number of Params of a call to a RPC Method is correct. If it finds a Method with the same name + number of params it checks if it is possible to cast all params given from the client to the needed one on server-side. If it cannot find the right type it will lookup the custom-mappings of the Adapter and call the function of the Type-Mapping.

The CastToObject Class is a more general approach to cast any given Table,Map to a JavaBean:
There is an exmple in the CustomMapping class that illustrates how to use it to directly invoke a Method with a type.


Release Notes:

Details:
    

Tests:



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@13070 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c