Change 20090602-maxcarlson-i by maxcarlson@Bank on 2009-06-02 16:37:12 PDT     in /Users/maxcarlson/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Fix IE to select inputtexts that share the same area as a text

Bugs Fixed: LPP-8219 - IE DHTML: inputtext focus/selection issues

Technical Reviewer: hminsky
QA Reviewer: mdemmon

Details: LzSprite - Add textgrabsinputtextfocus quirk, set to true for IE.

LzInputTextSprite - __hideIfNotFocused():  Shorten lookups to LzInputTextSprite.prototype.  Look at the srcElement of the last event to see if it was an instance of LzTextSprite - this happens when an text is in the same area as an inputtext.  If this is the case, focus the inputtext and avoid hiding the shown inputtext.

Tests: Testcase from LPP-8219 and compose window in webtop mail allows the to: field to be selected in the area that says 'Drag a contact or type for autosuggest'.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@14052 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c