Change 20090721-maxcarlson-x by maxcarlson@Bank.local on 2009-07-21 13:58:56 PDT     in /Users/maxcarlson/openlaszlo/trunk-clean     for http://svn.openlaszlo.org/openlaszlo/trunk
Summary: Prevent exceptions when hiding an inputtext that's not showing

Bugs Fixed: LPP-8301 - In dhtml an error occurs when mousing over an iframe when focus is in a text component that is not visible

Technical Reviewer: hminsky
QA Reviewer: mdemmon

Details: Check to see if the child is attached to its parent before attempting to remove it.

Tests: See LPP-8301.



git-svn-id: http://svn.openlaszlo.org/openlaszlo/trunk@14363 fa20e4f9-1d0a-0410-b5f3-dc9c16b8b17c